# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Solution",
        "description": "Description of solution API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/refunds": {
            "get": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Get list of refunds",
                "description": "Display a paginated list of refuns.",
                "operationId": "f6120e671b0e979a8458231e51021066",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of refunds per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of refunds"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Create a new refund",
                "description": "Store a newly created refund in storage.",
                "operationId": "4b30fa6a2d9e357459dd51684a31030a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refund_amount",
                                    "status",
                                    "transaction_id"
                                ],
                                "properties": {
                                    "refund_amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 150.75
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Pending"
                                    },
                                    "transaction_id": {
                                        "type": "integer",
                                        "example": 1023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Refund created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refunds/{id}": {
            "patch": {
                "tags": [
                    "Refunds"
                ],
                "summary": "Update an existing refund",
                "description": "Update an existing refund.",
                "operationId": "88f01f2225d4388d3e7acc10ce7f6ed6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Refund ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refund_amount",
                                    "status"
                                ],
                                "properties": {
                                    "refund_amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 200
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refund updated successfully"
                    },
                    "404": {
                        "description": "Refund not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get settings",
                "description": "Display the settings.",
                "operationId": "04b6f539c3af7d460d214177ac396d5e",
                "responses": {
                    "200": {
                        "description": "Settings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update settings",
                "description": "Update settings.",
                "operationId": "903d931bbeba813e2aef0b0ae75815fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "time_update_service",
                                    "time_update_order",
                                    "time_deny_order",
                                    "time_exchange_rate",
                                    "account_no"
                                ],
                                "properties": {
                                    "time_update_service": {
                                        "type": "string",
                                        "example": "14:30:00"
                                    },
                                    "time_update_order": {
                                        "type": "string",
                                        "example": "15:00:00"
                                    },
                                    "time_deny_order": {
                                        "type": "string",
                                        "example": "16:00:00"
                                    },
                                    "time_exchange_rate": {
                                        "type": "string",
                                        "example": "17:30:00"
                                    },
                                    "account_no": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123-456-7890"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "example": "https://facebook.com/example"
                                    },
                                    "zalo": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "bank_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}